select class deserves its own file
